# docs/openapi.yaml
openapi: 3.1.0
info:
  title: OpenAPI definition
  version: v0
servers:
- url: http://localhost
  description: Generated server url
tags:
- name: Agit
  description: 아지트 API
paths:
  /api/v1/agits:
    post:
      tags:
      - Agit
      summary: 아지트 생성
      description: 아지트와 방장 프로필을 함께 생성하고 초대 코드를 발급합니다.
      operationId: createAgit
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateAgitRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/CreateAgitResponse"
  /api/v1/agits/{code}/join:
    post:
      tags:
      - Agit
      summary: 아지트 입장
      description: "초대 코드로 아지트에 입장합니다. 신규는 GUEST 프로필을 생성하고, LEFT는 닉네임·이미지를 갱신한 뒤 ACTIVE로\
        \ 전환합니다. 이미 ACTIVE면 409, BANNED면 403, 정원 초과면 409를 반환합니다."
      operationId: joinAgit
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/JoinAgitRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "201":
          description: Created
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/JoinAgitResponse"
  /api/v1/agits/{agitUuid}/members/{userUuid}/unban:
    post:
      tags:
      - Agit
      summary: 아지트 밴 해제
      description: 아지트장이 userUuid 기준으로 밴을 해제합니다. BANNED인 멤버는 status를 LEFT로 바꾸고 bans
        이력에 해제 시각을 기록합니다. 이미 LEFT인 경우 변경 없이 성공합니다.
      operationId: unbanMember
      parameters:
      - name: agitUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: userUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorUserRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: No Content
  /api/v1/agits/{agitUuid}/members/{ampId}/transfer-host:
    post:
      tags:
      - Agit
      summary: 아지트장 위임
      description: 아지트장이 ampId의 ACTIVE 게스트에게 방장 권한을 위임합니다. 성공 시 204를 반환합니다.
      operationId: transferHost
      parameters:
      - name: agitUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: ampId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorUserRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: No Content
  /api/v1/agits/{agitUuid}/members/{ampId}/ban:
    post:
      tags:
      - Agit
      summary: 아지트에서 내보내기
      description: 아지트장이 ampId로 멤버를 내보냅니다. status를 BANNED로 바꾸고 bans 이력을 남깁니다. ACTIVE
        HOST는 내보낼 수 없습니다. 이미 BANNED인 경우 변경 없이 성공합니다.
      operationId: banMember
      parameters:
      - name: agitUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      - name: ampId
        in: path
        required: true
        schema:
          type: integer
          format: int64
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorUserRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: No Content
  /api/v1/agits/{agitUuid}/leave:
    post:
      tags:
      - Agit
      summary: 아지트 나가기
      description: "본인이 아지트에서 나갑니다. GUEST는 status를 LEFT로 변경합니다. HOST는 ACTIVE 인원이 본\
        인뿐일 때만 나갈 수 있으며, 이때 아지트는 소프트 삭제됩니다. 이미 LEFT이거나 BANNED인 경우 변경 없이 성공합니다."
      operationId: leaveAgit
      parameters:
      - name: agitUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorUserRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "204":
          description: No Content
  /api/v1/agits/{agitUuid}/invite-code:
    post:
      tags:
      - Agit
      summary: 초대 코드 재발급
      description: 아지트장이 초대 코드를 재발급합니다. 호출마다 새 코드를 반환합니다. 연타 방지는 FE에서 처리합니다.
      operationId: reissueInviteCode
      parameters:
      - name: agitUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorUserRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ReissueInviteCodeResponse"
  /api/v1/agits/{agitUuid}:
    patch:
      tags:
      - Agit
      summary: 아지트 정보 변경
      description: "아지트장이 제목·인원·소개글·섬네일을 수정합니다. 변경 인원수는 현재 ACTIVE 인원보다 작을 수 없으며, 그\
        \ 외 검증은 생성과 동일합니다."
      operationId: updateAgit
      parameters:
      - name: agitUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateAgitRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpdateAgitResponse"
  /api/v1/agits/{agitUuid}/members/me:
    patch:
      tags:
      - Agit
      summary: 내 아지트 프로필 수정
      description: 접속 유저가 해당 아지트에서 ACTIVE인 본인 닉네임·프로필 이미지를 부분 수정합니다. 미전달 필드는 유지합니다.
      operationId: updateMyMemberProfile
      parameters:
      - name: agitUuid
        in: path
        required: true
        schema:
          type: string
          format: uuid
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UpdateMyMemberProfileRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/UpdateMyMemberProfileResponse"
  /api/v1/agits/{code}/landing:
    get:
      tags:
      - Agit
      summary: 아지트 랜딩 조회
      description: 초대 코드로 ACTIVE 아지트의 랜딩 표시 정보를 조회합니다. 인증 없이 호출 가능합니다.
      operationId: getLandingByCode
      parameters:
      - name: code
        in: path
        required: true
        schema:
          type: string
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/AgitLandingResponse"
  /api/v1/agits/me:
    get:
      tags:
      - Agit
      summary: 내 아지트 목록 조회
      description: "접속 유저가 ACTIVE로 속한 아지트 목록(제목·UUID)을 반환합니다. 정렬은 agit.updated_at\
        \ 내림차순(임시)이며, 이후 최근 토픽순으로 교체 예정입니다. userUuid는 임시로 request body로 받습니다."
      operationId: listMyAgits
      requestBody:
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ActorUserRequest"
        required: true
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/MyAgitItemResponse"
  /api/test:
    get:
      tags:
      - test-controller
      operationId: test
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
  /api/test/ping:
    get:
      tags:
      - test-controller
      operationId: ping
      responses:
        "409":
          description: Conflict
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "404":
          description: Not Found
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "403":
          description: Forbidden
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "400":
          description: Bad Request
          content:
            '*/*':
              schema:
                $ref: "#/components/schemas/ErrorResponse"
        "200":
          description: OK
          content:
            '*/*':
              schema:
                type: string
components:
  schemas:
    ErrorResponse:
      type: object
      description: 에러 응답
      properties:
        code:
          type: string
          description: 에러 코드
          example: ALREADY_JOINED
        message:
          type: string
          description: 에러 메시지
          example: 이미 참여 중인 아지트입니다.
        ampId:
          type: integer
          format: int64
          description: 이미 가입된 멤버 프로필 ID (해당 시)
          example: 1
    CreateAgitRequest:
      type: object
      description: 아지트 생성 요청
      properties:
        agitName:
          type: string
          description: "아지트 제목 (필수, 최대 20자)"
          example: 주말 보드게임
        description:
          type: string
          description: "소개글 (선택, 최대 100자)"
          example: 가볍게 즐겨요
        maximumCapacity:
          type: integer
          format: int32
          description: "최대 인원 (필수, 기본 허용 상한 5)"
          example: 5
        thumbnailPath:
          type: string
          description: "섬네일 경로 (선택, 최대 255자)"
          example: agits/thumbnails/sample.png
        userUuid:
          type: string
          format: uuid
          description: "방장 사용자 UUID (UUIDv7, 임시 body 전달 — 추후 인증에서 추출)"
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
        nickname:
          type: string
          description: "방장 닉네임 (필수, 영문·숫자·한글 2~12자, 특수문자·공백 불가)"
          example: 보드왕
        profileImagePath:
          type: string
          description: 방장 프로필 이미지 경로 (선택)
          example: profiles/host.png
    CreateAgitResponse:
      type: object
      description: 아지트 생성 응답
      properties:
        agitUuid:
          type: string
          format: uuid
          description: 아지트 UUID
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
        agitName:
          type: string
          description: 아지트 제목
          example: 주말 보드게임
        description:
          type: string
          description: 소개글
          example: 가볍게 즐겨요
        maximumCapacity:
          type: integer
          format: int32
          description: 최대 인원
          example: 5
        code:
          type: string
          description: 초대 코드
          example: AB12CD
        thumbnailPath:
          type: string
          description: 섬네일 경로
          example: agits/thumbnails/sample.png
        ampId:
          type: integer
          format: int64
          description: 방장 프로필 ID (amp_id)
          example: 1
        nickname:
          type: string
          description: 방장 닉네임
          example: 보드왕
        profileImagePath:
          type: string
          description: 방장 프로필 이미지 경로
          example: profiles/host.png
        role:
          type: string
          description: 방장 역할
          enum:
          - HOST
          - GUEST
          example: HOST
    JoinAgitRequest:
      type: object
      description: 아지트 입장 요청
      properties:
        userUuid:
          type: string
          format: uuid
          description: "사용자 UUID (UUIDv7, 임시 body 전달 — 추후 인증에서 추출)"
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
        nickname:
          type: string
          description: "닉네임 (필수, 영문·숫자·한글 2~12자, 특수문자·공백 불가)"
          example: 보드왕
        profileImagePath:
          type: string
          description: 프로필 이미지 경로 (선택)
          example: profiles/guest.png
    JoinAgitResponse:
      type: object
      description: 아지트 입장 응답
      properties:
        agitUuid:
          type: string
          format: uuid
          description: 아지트 UUID
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
        ampId:
          type: integer
          format: int64
          description: 멤버 프로필 ID (amp_id)
          example: 1
        nickname:
          type: string
          description: 닉네임
          example: 보드왕
        profileImagePath:
          type: string
          description: 프로필 이미지 경로
          example: profiles/guest.png
        role:
          type: string
          description: 역할
          enum:
          - HOST
          - GUEST
          example: GUEST
    ActorUserRequest:
      type: object
      description: 액터 사용자 UUID 요청 (임시 — 인증 연동 전)
      properties:
        userUuid:
          type: string
          format: uuid
          description: "요청자 사용자 UUID (UUIDv7, 임시 body 전달 — 추후 인증에서 추출)"
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
    ReissueInviteCodeResponse:
      type: object
      description: 초대 코드 재발급 응답
      properties:
        code:
          type: string
          description: 새로 발급된 초대 코드
          example: AB12CD
    UpdateAgitRequest:
      type: object
      description: 아지트 정보 변경 요청
      properties:
        userUuid:
          type: string
          format: uuid
          description: "요청자 사용자 UUID (UUIDv7, 임시 body 전달 — 추후 인증에서 추출)"
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
        agitName:
          type: string
          description: "아지트 제목 (필수, 최대 20자)"
          example: 주말 보드게임
        description:
          type: string
          description: "소개글 (선택, 최대 100자)"
          example: 가볍게 즐겨요
        maximumCapacity:
          type: integer
          format: int32
          description: "최대 인원 (필수, 기본 허용 상한 5, 현재 ACTIVE 인원 이상)"
          example: 5
        thumbnailPath:
          type: string
          description: "섬네일 경로 (선택, 최대 255자)"
          example: agits/thumbnails/sample.png
    UpdateAgitResponse:
      type: object
      description: 아지트 정보 변경 응답
      properties:
        agitUuid:
          type: string
          format: uuid
          description: 아지트 UUID
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
        agitName:
          type: string
          description: 아지트 제목
          example: 주말 보드게임
        description:
          type: string
          description: 소개글
          example: 가볍게 즐겨요
        maximumCapacity:
          type: integer
          format: int32
          description: 최대 인원
          example: 5
        thumbnailPath:
          type: string
          description: 섬네일 경로
          example: agits/thumbnails/sample.png
    UpdateMyMemberProfileRequest:
      type: object
      description: 내 아지트 멤버 프로필 수정 요청
      properties:
        userUuid:
          type: string
          format: uuid
          description: "요청자 사용자 UUID (UUIDv7, 임시 body 전달 — 추후 인증에서 추출)"
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
        nickname:
          type: string
          description: "닉네임 (선택, 영문·숫자·한글 2~12자, 특수문자·공백 불가). 미전달 시 유지"
          example: 보드왕
        profileImagePath:
          type: string
          description: "프로필 이미지 경로 (선택). 미전달 시 유지, 빈 문자열이면 제거"
          example: profiles/guest.png
    UpdateMyMemberProfileResponse:
      type: object
      description: 내 아지트 멤버 프로필 수정 응답
      properties:
        nickname:
          type: string
          description: 닉네임
          example: 보드왕
        profileImagePath:
          type: string
          description: 프로필 이미지 경로
          example: profiles/guest.png
    AgitLandingResponse:
      type: object
      description: 아지트 랜딩 조회 응답 (표시 전용)
      properties:
        agitName:
          type: string
          description: 아지트 제목
          example: 주말 보드게임
        description:
          type: string
          description: 소개글
          example: 가볍게 즐겨요
        currentMemberCount:
          type: integer
          format: int64
          description: 현재 인원 (ACTIVE 멤버 수)
          example: 1
        maximumCapacity:
          type: integer
          format: int32
          description: 최대 인원
          example: 5
        hostNickname:
          type: string
          description: 아지트장 닉네임
          example: 보드왕
        thumbnailPath:
          type: string
          description: 섬네일 경로
          example: agits/thumbnails/sample.png
    MyAgitItemResponse:
      type: object
      description: 내 아지트 목록 항목
      properties:
        agitUuid:
          type: string
          format: uuid
          description: 아지트 UUID
          example: 018f3f6e-8e2a-7b3c-9d4e-5f6a7b8c9d0e
        agitName:
          type: string
          description: 아지트 제목
          example: 주말 보드게임