# src/main/java/com/plip/agit/application/port/in/dto/UpdateMyMemberProfileRequestDto.java
package com.plip.agit.application.port.in.dto;

import java.util.UUID;
import lombok.Builder;
import lombok.Getter;

@Getter
@Builder
public class UpdateMyMemberProfileRequestDto {

	private final UUID userUuid;
	private final String nickname;
	private final String profileImagePath;
}